Add Labels reducer file to reducers
This does contain all reducers and parsers for the
labels.
