+++ plugins/services/src/js/components/forms/EnvironmentFormSection.js
@@ -4,10 +4,9 @@ import FieldError from '../../../../../../src/js/components/form/FieldError';
 import FieldInput from '../../../../../../src/js/components/form/FieldInput';
 import FieldLabel from '../../../../../../src/js/components/form/FieldLabel';
 import FormGroup from '../../../../../../src/js/components/form/FormGroup';
-import TransactionTypes from '../../../../../../src/js/constants/TransactionTypes';
 import ReducerUtil from '../../../../../../src/js/utils/ReducerUtil';
-import Util from '../../../../../../src/js/utils/Util';
 import {FormReducer as env} from '../../reducers/form/EnvironmentVariables';
+import {FormReducer as labels} from '../../reducers/form/Labels';
 
 class EnvironmentFormSection extends Component {
   constructor() {
@@ -206,51 +205,7 @@ EnvironmentFormSection.propTypes = {
 
 EnvironmentFormSection.configReducers = {
   env,
-  labels(state = [], {type, path, value}) {
-    // Prepare
-    // TODO: Remove when we use the parsers
-    if (Util.isObject(state) && !Array.isArray(state)) {
-      state = Object.keys(state).reduce((memo, key) => {
-        memo.push({
-          key,
-          value: state[key]
-        });
-        return memo;
-      }, []);
-    }
-
-    // ROWS
-    if (path != null && path.join('.').search('labels') !== -1) {
-      if (path.join('.') === 'labels') {
-        switch (type) {
-          case TransactionTypes.ADD_ROW:
-            state.push({key: null, value: null});
-            break;
-          case TransactionTypes.REMOVE_ROW:
-            state = state.filter((item, index) => {
-              return index !== value;
-            });
-            break;
-        }
-        return state;
-      }
-
-      // SET
-      let joinedPath = path.join('.');
-      let index = joinedPath.match(/\d+/)[0];
-      switch (true) {
-        case (type === TransactionTypes.SET &&
-        `labels.${index}.key` === joinedPath):
-          state[index].key = value;
-          break;
-        case (type === TransactionTypes.SET &&
-        `labels.${index}.value` === joinedPath):
-          state[index].value = value;
-          break;
-      }
-    }
-    return state;
-  }
+  labels
 };
 
 EnvironmentFormSection.validationReducers = {
